Update2

+++ MIATZ/Models/Patient.cs
@@ -14,8 +14,8 @@ namespace MIATZ.Models
         public string Surname { get; set; } //фамилия
         public string Patronymic { get; set; } //отчество
         public DateTime DateOfBirthDay { get; set; } // возраст
-        public int Height { get; set; }// рост
-        public int Weight { get; set; }//вес
+        public double Height { get; set; }// рост
+        public double Weight { get; set; }//вес
 
     }
 }
